# internal/server/server.go
package server

import (
	"context"
	"crypto/subtle"
	"encoding/json"
	"fmt"
	"net"
	"net/http"
	"net/http/pprof"
	"strings"
	"sync"
	"time"

	"github.com/DocHoax/watchdog/internal/alerts"
	"github.com/DocHoax/watchdog/internal/anomaly"
	"github.com/DocHoax/watchdog/internal/audit"
	"github.com/DocHoax/watchdog/internal/collector"
	"github.com/DocHoax/watchdog/internal/config"
	"github.com/DocHoax/watchdog/internal/diagnostics"
	"github.com/DocHoax/watchdog/internal/fleet"
	"github.com/DocHoax/watchdog/internal/logger"
	"github.com/DocHoax/watchdog/internal/storage"
	"github.com/DocHoax/watchdog/pkg/model"
)

// IsLoopback reports whether the given bind address is a loopback/local address.
// The addresses 127.0.0.1, ::1, and "localhost" are treated as loopback.
// An empty string is also considered loopback because it falls back to 127.0.0.1.
func IsLoopback(addr string) bool {
	if addr == "" || strings.EqualFold(addr, "localhost") {
		return true
	}
	ip := net.ParseIP(addr)
	if ip == nil {
		return false
	}
	return ip.IsLoopback()
}

// ValidateServerSecurity checks that the server configuration meets security
// requirements for the configured bind address. Non-loopback addresses require
// both an authentication token and a complete TLS configuration. Incomplete TLS
// configuration (cert without key or vice versa) is always rejected regardless
// of bind address.
func ValidateServerSecurity(cfg *config.AgentConfig) error {
	// Reject incomplete TLS configuration regardless of bind address.
	hasCert := cfg.TLSCert != ""
	hasKey := cfg.TLSKey != ""
	if hasCert != hasKey {
		if hasCert {
			return fmt.Errorf("incomplete TLS configuration: tls_cert is set but tls_key is missing; provide both or neither")
		}
		return fmt.Errorf("incomplete TLS configuration: tls_key is set but tls_cert is missing; provide both or neither")
	}

	bindAddr := cfg.BindAddress
	if IsLoopback(bindAddr) {
		// Loopback binding is safe without token or TLS.
		return nil
	}

	// Non-loopback binding requires authentication.
	if cfg.Token == "" {
		return fmt.Errorf(
			"Watchdog server refuses to bind to %s without authentication: "+
				"externally exposed API requires a token; configure agent.token "+
				"or bind to localhost (127.0.0.1)",
			bindAddr,
		)
	}

	// Non-loopback binding requires TLS.
	if !hasCert {
		return fmt.Errorf(
			"Watchdog server refuses to bind to %s without TLS: "+
				"externally exposed API requires encryption; configure agent.tls_cert "+
				"and agent.tls_key or bind to localhost (127.0.0.1)",
			bindAddr,
		)
	}

	return nil
}

// Server represents the Watchdog HTTP API server / remote agent.
type Server struct {
	cfg       *config.Config
	collector *collector.Manager
	storage   storage.Storage
	diagEng   *diagnostics.Engine
	alertEng  *alerts.Engine
	anomDet   *anomaly.Detector
	exporter     *PrometheusExporter
	auditLog     audit.AuditLogger
	fleetService fleet.FleetService

	httpServer *http.Server
	startTime  time.Time

	mu             sync.RWMutex
	latestSnapshot *model.SystemSnapshot
	latestDiag     *model.DiagnosticReport
	activeAlerts   []model.AlertEvent
	latestAnomaly  *model.AnomalyReport
}

// NewServer initializes a new Server.
func NewServer(
	cfg *config.Config,
	col *collector.Manager,
	store storage.Storage,
	diagEng *diagnostics.Engine,
	alertEng *alerts.Engine,
	anomDet *anomaly.Detector,
) *Server {
	var auditLog audit.AuditLogger
	if cfg != nil && cfg.Audit.Enabled {
		auditLog = audit.New(cfg.Audit, store, nil)
	} else {
		auditLog = audit.NewNopAuditLogger()
	}

	var fleetSvc fleet.FleetService
	if store != nil && cfg != nil {
		fleetSvc = fleet.NewFleetService(store, cfg.Fleet)
	}

	return &Server{
		cfg:          cfg,
		collector:    col,
		storage:      store,
		diagEng:      diagEng,
		alertEng:     alertEng,
		anomDet:      anomDet,
		exporter:     NewPrometheusExporter(),
		auditLog:     auditLog,
		fleetService: fleetSvc,
		startTime:    time.Now(),
	}
}

// SetAuditLogger overrides or configures the audit logger for the server.
func (s *Server) SetAuditLogger(al audit.AuditLogger) {
	s.mu.Lock()
	defer s.mu.Unlock()
	s.auditLog = al
}

// SetFleetService configures the fleet service instance for the server.
func (s *Server) SetFleetService(fs fleet.FleetService) {
	s.mu.Lock()
	defer s.mu.Unlock()
	s.fleetService = fs
}

// FleetService returns the configured fleet service instance.
func (s *Server) FleetService() fleet.FleetService {
	s.mu.RLock()
	defer s.mu.RUnlock()
	return s.fleetService
}

// Start runs the HTTP server and background collection worker until ctx is cancelled.
func (s *Server) Start(ctx context.Context) error {
	// Validate security requirements before starting.
	if err := ValidateServerSecurity(&s.cfg.Agent); err != nil {
		if s.auditLog != nil {
			_ = s.auditLog.Record(context.Background(), model.AuditEvent{
				ID:        audit.GenerateEventID(),
				Timestamp: time.Now().UTC(),
				EventType: model.EventServerStartFailure,
				Severity:  model.AuditSeverityCritical,
				Outcome:   model.AuditOutcomeFailure,
				Actor: model.AuditActor{
					Type:     model.ActorTypeSystem,
					Identity: "watchdog-daemon",
				},
				Source: model.AuditSource{
					Address: s.cfg.Agent.BindAddress,
				},
				Message: fmt.Sprintf("Server security validation failed: %v", err),
			})
		}
		return fmt.Errorf("server security check failed: %w", err)
	}

	mux := http.NewServeMux()

	// Register routes
	mux.HandleFunc("/health", s.handleHealth)
	mux.HandleFunc("/healthz", s.handleHealth)
	mux.HandleFunc("/api/v1/health", s.handleHealth)
	mux.HandleFunc("/ready", s.handleReady)
	mux.HandleFunc("/readyz", s.handleReady)
	mux.HandleFunc("/api/v1/ready", s.handleReady)
	mux.HandleFunc("/metrics", s.exporter.Handler())

	// Secured API routes — authentication is always enforced when a token is
	// configured; when binding to a non-loopback address the validation above
	// guarantees a token is present.
	mux.Handle("/api/v1/snapshot", s.authMiddleware(http.HandlerFunc(s.handleSnapshot)))
	mux.Handle("/api/v1/diagnostics", s.authMiddleware(http.HandlerFunc(s.handleDiagnostics)))
	mux.Handle("/api/v1/alerts", s.authMiddleware(http.HandlerFunc(s.handleAlerts)))
	mux.Handle("/api/v1/anomalies", s.authMiddleware(http.HandlerFunc(s.handleAnomalies)))
	mux.Handle("/api/v1/audit/events", s.authMiddleware(http.HandlerFunc(s.handleAuditEvents)))

	// Fleet API routes
	mux.Handle("/api/v1/node", s.authMiddleware(http.HandlerFunc(s.handleNodeIdentity)))
	mux.Handle("/api/v1/heartbeat", s.authMiddleware(http.HandlerFunc(s.handleHeartbeat)))
	mux.Handle("/api/v1/telemetry", s.authMiddleware(http.HandlerFunc(s.handleTelemetry)))
	mux.Handle("/api/v1/fleet", s.authMiddleware(http.HandlerFunc(s.handleFleetRoute)))
	mux.Handle("/api/v1/fleet/", s.authMiddleware(http.HandlerFunc(s.handleFleetRoute)))

	// Diagnostic Profiling (pprof)
	mux.HandleFunc("/debug/pprof/", pprof.Index)
	mux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
	mux.HandleFunc("/debug/pprof/profile", pprof.Profile)
	mux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
	mux.HandleFunc("/debug/pprof/trace", pprof.Trace)
	mux.Handle("/debug/pprof/heap", pprof.Handler("heap"))
	mux.Handle("/debug/pprof/goroutine", pprof.Handler("goroutine"))
	mux.Handle("/debug/pprof/allocs", pprof.Handler("allocs"))
	mux.Handle("/debug/pprof/block", pprof.Handler("block"))
	mux.Handle("/debug/pprof/mutex", pprof.Handler("mutex"))

	bindAddr := s.cfg.Agent.BindAddress
	if bindAddr == "" {
		bindAddr = "127.0.0.1"
	}
	port := s.cfg.Agent.Port
	if port == 0 {
		port = 8443
	}

	addr := fmt.Sprintf("%s:%d", bindAddr, port)
	handler := RequestIDMiddleware(
		PanicRecoveryMiddleware(
			MaxBodySizeMiddleware(DefaultMaxRequestBodySize)(
				RequestLoggerMiddleware(mux),
			),
			s.auditLog,
		),
	)

	s.httpServer = &http.Server{
		Addr:              addr,
		Handler:           handler,
		ReadHeaderTimeout: 10 * time.Second,
		ReadTimeout:       30 * time.Second,
		WriteTimeout:      30 * time.Second,
		IdleTimeout:       60 * time.Second,
	}

	// Start background collection loop
	go s.runCollectionLoop(ctx)

	// Channel for server errors
	errChan := make(chan error, 1)

	// Record server lifecycle audit events
	if s.auditLog != nil {
		_ = s.auditLog.Record(ctx, model.AuditEvent{
			ID:        audit.GenerateEventID(),
			Timestamp: time.Now().UTC(),
			EventType: model.EventServerStart,
			Severity:  model.AuditSeverityInfo,
			Outcome:   model.AuditOutcomeSuccess,
			Actor: model.AuditActor{
				Type:     model.ActorTypeSystem,
				Identity: "watchdog-daemon",
			},
			Source: model.AuditSource{
				Address: addr,
			},
			Message: fmt.Sprintf("Watchdog API / Agent server starting on %s", addr),
			Metadata: map[string]string{
				"bind_address": bindAddr,
				"port":         fmt.Sprintf("%d", port),
				"tls_enabled":  fmt.Sprintf("%t", s.cfg.Agent.TLSCert != ""),
			},
		})

		tlsEventType := model.EventTLSDisabled
		tlsMsg := "TLS encryption disabled for server (loopback binding)"
		if s.cfg.Agent.TLSCert != "" && s.cfg.Agent.TLSKey != "" {
			tlsEventType = model.EventTLSEnabled
			tlsMsg = "TLS encryption enabled for server"
		}
		_ = s.auditLog.Record(ctx, model.AuditEvent{
			ID:        audit.GenerateEventID(),
			Timestamp: time.Now().UTC(),
			EventType: tlsEventType,
			Severity:  model.AuditSeverityInfo,
			Outcome:   model.AuditOutcomeSuccess,
			Actor: model.AuditActor{
				Type:     model.ActorTypeSystem,
				Identity: "watchdog-daemon",
			},
			Source: model.AuditSource{
				Address: addr,
			},
			Message: tlsMsg,
		})
	}

	go func() {
		logger.Infof("Watchdog API / Agent server listening on %s (TLS: %v)", addr, s.cfg.Agent.TLSCert != "")
		if s.cfg.Agent.TLSCert != "" && s.cfg.Agent.TLSKey != "" {
			if err := s.httpServer.ListenAndServeTLS(s.cfg.Agent.TLSCert, s.cfg.Agent.TLSKey); err != nil && err != http.ErrServerClosed {
				errChan <- err
			}
		} else {
			if err := s.httpServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
				errChan <- err
			}
		}
	}()

	// Wait for context cancellation or server error
	select {
	case <-ctx.Done():
		logger.Infof("Shutting down Watchdog server...")
		if s.auditLog != nil {
			_ = s.auditLog.Record(context.Background(), model.AuditEvent{
				ID:        audit.GenerateEventID(),
				Timestamp: time.Now().UTC(),
				EventType: model.EventServerStop,
				Severity:  model.AuditSeverityInfo,
				Outcome:   model.AuditOutcomeSuccess,
				Actor: model.AuditActor{
					Type:     model.ActorTypeSystem,
					Identity: "watchdog-daemon",
				},
				Source: model.AuditSource{
					Address: addr,
				},
				Message: "Watchdog API / Agent server stopped",
			})
		}
		shutdownCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
		defer cancel()
		return s.httpServer.Shutdown(shutdownCtx)
	case err := <-errChan:
		if s.auditLog != nil {
			_ = s.auditLog.Record(context.Background(), model.AuditEvent{
				ID:        audit.GenerateEventID(),
				Timestamp: time.Now().UTC(),
				EventType: model.EventServerStartFailure,
				Severity:  model.AuditSeverityCritical,
				Outcome:   model.AuditOutcomeFailure,
				Actor: model.AuditActor{
					Type:     model.ActorTypeSystem,
					Identity: "watchdog-daemon",
				},
				Source: model.AuditSource{
					Address: addr,
				},
				Message: fmt.Sprintf("Server execution failed: %v", err),
			})
		}
		return err
	}
}

// runCollectionLoop periodically collects system metrics, evaluates alerts, and updates the exporter.
func (s *Server) runCollectionLoop(ctx context.Context) {
	interval := s.cfg.RefreshInterval
	if interval < 500*time.Millisecond {
		interval = 1 * time.Second
	}

	ticker := time.NewTicker(interval)
	defer ticker.Stop()

	// Initial collection
	s.collectAndEvaluate(ctx)

	for {
		select {
		case <-ctx.Done():
			return
		case <-ticker.C:
			s.collectAndEvaluate(ctx)
		}
	}
}

func (s *Server) collectAndEvaluate(ctx context.Context) {
	if s.collector == nil {
		return
	}

	snap, err := s.collector.CollectAll(ctx)
	if err != nil {
		logger.Warnf("Collector error in server loop: %v", err)
		return
	}

	var diag *model.DiagnosticReport
	if s.diagEng != nil {
		diag, _ = s.diagEng.Run(ctx, snap)
	}

	var activeAlerts []model.AlertEvent
	if s.alertEng != nil {
		_, _, _ = s.alertEng.Evaluate(ctx, snap)
		activeAlerts = s.alertEng.GetActiveAlerts()
	}

	var anom *model.AnomalyReport
	if s.anomDet != nil {
		anom = s.anomDet.FeedSnapshot(snap)
	}

	// Update storage if configured
	if s.storage != nil {
		_ = s.storage.SaveSnapshot(ctx, snap)
	}

	// Update in-memory state
	s.mu.Lock()
	s.latestSnapshot = snap
	s.latestDiag = diag
	s.activeAlerts = activeAlerts
	s.latestAnomaly = anom
	s.mu.Unlock()

	// Update Prometheus exporter
	s.exporter.Update(snap, diag, activeAlerts, anom)
}

// authMiddleware enforces bearer token authentication if configured and logs audit events.
func (s *Server) authMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		token := s.cfg.Agent.Token
		reqID := GetRequestID(r.Context())

		if token != "" {
			authHeader := r.Header.Get("Authorization")
			customHeader := r.Header.Get("X-Watchdog-Token")

			provided := ""
			if len(authHeader) > 7 && strings.EqualFold(authHeader[:7], "Bearer ") {
				provided = strings.TrimSpace(authHeader[7:])
			} else if customHeader != "" {
				provided = strings.TrimSpace(customHeader)
			}

			if provided == "" {
				if s.auditLog != nil {
					_ = s.auditLog.Record(r.Context(), model.AuditEvent{
						ID:        audit.GenerateEventID(),
						Timestamp: time.Now().UTC(),
						EventType: model.EventAuthMissingCredentials,
						Severity:  model.AuditSeverityWarning,
						Outcome:   model.AuditOutcomeDenied,
						Actor: model.AuditActor{
							Type:     model.ActorTypeAnonymousClient,
							Identity: "anonymous",
						},
						Source: model.AuditSource{
							Address:   r.RemoteAddr,
							Endpoint:  r.URL.Path,
							Method:    r.Method,
							RequestID: reqID,
							UserAgent: r.UserAgent(),
						},
						Message: "Authentication failed: missing credentials",
					})
				}
				s.writeJSONError(w, http.StatusUnauthorized, "Unauthorized: missing authentication token")
				return
			}

			if subtle.ConstantTimeCompare([]byte(provided), []byte(token)) != 1 {
				// Audit with masked token (never raw token or authorization header)
				if s.auditLog != nil {
					_ = s.auditLog.Record(r.Context(), model.AuditEvent{
						ID:        audit.GenerateEventID(),
						Timestamp: time.Now().UTC(),
						EventType: model.EventAuthInvalidCredentials,
						Severity:  model.AuditSeverityWarning,
						Outcome:   model.AuditOutcomeDenied,
						Actor: model.AuditActor{
							Type:     model.ActorTypeAnonymousClient,
							Identity: audit.MaskToken(provided),
						},
						Source: model.AuditSource{
							Address:   r.RemoteAddr,
							Endpoint:  r.URL.Path,
							Method:    r.Method,
							RequestID: reqID,
							UserAgent: r.UserAgent(),
						},
						Message: "Authentication failed: invalid credentials",
					})
				}
				s.writeJSONError(w, http.StatusUnauthorized, "Unauthorized: invalid authentication token")
				return
			}

			// Successful authentication
			if s.auditLog != nil {
				_ = s.auditLog.Record(r.Context(), model.AuditEvent{
					ID:        audit.GenerateEventID(),
					Timestamp: time.Now().UTC(),
					EventType: model.EventAuthSuccess,
					Severity:  model.AuditSeverityInfo,
					Outcome:   model.AuditOutcomeSuccess,
					Actor: model.AuditActor{
						Type:     model.ActorTypeAuthenticatedClient,
						Identity: "api-client",
					},
					Source: model.AuditSource{
						Address:   r.RemoteAddr,
						Endpoint:  r.URL.Path,
						Method:    r.Method,
						RequestID: reqID,
						UserAgent: r.UserAgent(),
					},
					Message: "Authentication succeeded",
				})
			}
		}
		next.ServeHTTP(w, r)
	})
}

func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}
	uptime := time.Since(s.startTime).Seconds()
	resp := map[string]any{
		"status":         "ok",
		"uptime_seconds": uptime,
		"timestamp":      time.Now().UTC().Format(time.RFC3339),
		"version":        "1.0.0",
	}
	s.writeJSON(w, http.StatusOK, resp)
}

func (s *Server) handleReady(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}

	s.mu.RLock()
	snap := s.latestSnapshot
	s.mu.RUnlock()

	ready := true
	collectorsStatus := "ok"
	if snap == nil {
		ready = false
		collectorsStatus = "waiting_for_initial_snapshot"
	}

	storageStatus := "disabled"
	if s.storage != nil {
		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
		defer cancel()
		if err := s.storage.Ping(ctx); err != nil {
			ready = false
			storageStatus = fmt.Sprintf("error: %v", err)
		} else {
			storageStatus = "ok"
		}
	}

	diagnosticsStatus := "disabled"
	if s.diagEng != nil {
		diagnosticsStatus = "ok"
	}

	statusStr := "ready"
	httpStatus := http.StatusOK
	if !ready {
		statusStr = "not_ready"
		httpStatus = http.StatusServiceUnavailable
	}

	resp := map[string]any{
		"status":      statusStr,
		"collectors":  collectorsStatus,
		"diagnostics": diagnosticsStatus,
		"storage":     storageStatus,
		"timestamp":   time.Now().UTC().Format(time.RFC3339),
		"version":     "1.0.0",
	}

	s.writeJSON(w, httpStatus, resp)
}

func (s *Server) handleSnapshot(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}
	s.mu.RLock()
	snap := s.latestSnapshot
	s.mu.RUnlock()

	if snap == nil {
		s.writeJSONError(w, http.StatusServiceUnavailable, "System metrics snapshot not yet available")
		return
	}

	s.writeJSON(w, http.StatusOK, snap)
}

func (s *Server) handleDiagnostics(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}
	s.mu.RLock()
	diag := s.latestDiag
	s.mu.RUnlock()

	if diag == nil {
		s.writeJSONError(w, http.StatusServiceUnavailable, "Diagnostic report not yet available")
		return
	}

	s.writeJSON(w, http.StatusOK, diag)
}

func (s *Server) handleAlerts(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}
	s.mu.RLock()
	alerts := s.activeAlerts
	s.mu.RUnlock()

	s.writeJSON(w, http.StatusOK, map[string]any{
		"active_alerts": alerts,
		"count":         len(alerts),
		"timestamp":     time.Now().UTC().Format(time.RFC3339),
	})
}

func (s *Server) handleAnomalies(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet && r.Method != http.MethodHead {
		w.Header().Set("Allow", "GET, HEAD")
		s.writeJSONError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}
	s.mu.RLock()
	anom := s.latestAnomaly
	s.mu.RUnlock()

	if anom == nil {
		s.writeJSONError(w, http.StatusServiceUnavailable, "Anomaly detection report not yet available")
		return
	}

	s.writeJSON(w, http.StatusOK, anom)
}

func (s *Server) writeJSON(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(data)
}

func (s *Server) writeJSONError(w http.ResponseWriter, status int, message string) {
	s.writeJSON(w, status, map[string]string{
		"error": message,
	})
}